Move `Selection` to its own header.
The new header is `<bbp/sonata/selection.h>` and it's included in
`<bbp/sonata/population.h`. This ensures doesn't affect which include provide
`Selection` in an adverse way.
